requirements.txt
camel-ai[all]==0.2.43
chunkr-ai>=0.0.41
docx2markdown>=0.1.1
gradio>=3.50.2
